[GH-1398]: Fix parsing of numeric input value when generating Contributions shortcode.
When generating Contribution, JS goes through each input and runs `parseFloat` check on the value. This is incorrect since it allows '2,00' value to pass the `parseFloat( '2,00' ) > 0.00` check and then multiplying this string by 100 results in NaN and incorrect value in shortcode.

In a fix, the input value is parsed as integer and then converted to float using `toFixed( 2 )`. This ensures that the check is correct and multiply operation will return int.

Ref: https://github.com/laterpay/laterpay-wordpress-plugin/issues/1398